Combine all kernel module packages into one rpm package
When the kernel modules are splited into the seperated rpm package,
it will waste too time to upgrade the system kernel. It is a bad OOBE
for user.
So we supply one rpm package which includes all kernel modules, and it
could upgrade the kernel modules in a short time.

Signed-off-by: Guojian Zhou <guojian.zhou@windriver.com>