perl: Combine all perl module packages into one package
When the perl modules are splited into the individual package, it will
waste
too time to upgrade the system perl modules. It is a bad OOBE for user.

So we supply one rpm package which includes all perl modules, and it
could upgrade the perl modules in a short time.

Add the "PERL_MODULE_COMBINED_PACKAGE" flag to control whether create the
single large meta-package or inhibit the individual ones by default.

Signed-off-by: Guojian Zhou <guojian.zhou@windriver.com>