avahi: Add a new package for libdns_sd
When the libavahi-core package include libavahi-core.so and libdns_sd,
but they are not the same version share lib, such as libavahi-core.so.7
and libdns_sd.so.1, then the libavahi-core package will be named as
"libavahi-core" and not the expected name "libavahi-core7".
It will cause the released Pulsar system upgrade failure on the file
conflict of "libavahi-core.so.7.0.2". Because the system will find the
"libavahi-core7.rpm" and new "libavahi-core.rpm" supply the same name
share lib "libavahi-core.so.7.0.2".

So we split a new package "libdns_sd" to avoid this issue.

Signed-off-by: Guojian Zhou <guojian.zhou@windriver.com>