lxc: match base recipe postinst key name
We are currently getting the following build warnings as a result of
the changes made in commit 775fb905602a65f0368aad8eee5579c5d7cadbef
[wrlab: Fix some recipes upgrade exexute populate-volatile.sh issue]:

  WARNING: Variable key pkg_postinst_${PN} (
  if [ -z "$D" ] && [ -e /etc/init.d/populate-volatile.sh ] ; then
    /etc/init.d/populate-volatile.sh update
  fi
  ) replaces original key pkg_postinst_lxc (
  readlink -f /sbin/init | grep systemd && exit 0
  ).

Inspecting the rpm (rpm --scripts -qp) we can determine that his is
truely just a warning as the script is properly formed. But by using
'${PN}' instead of 'lxc' we can get the same result, without the
warnings. Again inspecting the rpm scripts confirms this. Fixes #56.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Jason Wessel <jason.wessel@windriver.com>