split /boot into a seperated partition
This is the preparation for encrypted rootfs, so we split /boot into an
unecnrypted partition.

As ostree needs to create symlink on boot partition, and vfat does not
support symlink, so we can't move /boot into ESP. That's why we create
this boot partition, and it will be labeled as "otaboot".

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>